+++ src-tauri/src/lib.rs
@@ -346,6 +346,7 @@ pub fn run() {
         .plugin(tauri_plugin_window_state::Builder::default().build())
         .plugin(tauri_plugin_updater::Builder::new().build())
         .plugin(tauri_plugin_process::init())
+        .plugin(tauri_plugin_clipboard_manager::init())
         .manage(pty_state.clone())
         .manage(pty_session_count)
         .manage(watcher_state)